Ensure failures are detected by Syncthing
If there is an unexpected problem, we need to relay that through
Syncthing back to the user. If we fail to commit, we do exit non-zero.
But Syncthing seems to ignore this and continue anyway. To get Syncthing
to fail, we also need to ensure that the original file still exists.